requirements.txt
django==2.2.17
celery==4.2.1
djangorestframework==3.12.2
gunicorn==20.0.4
requests==2.25.1
